+++ app/scripts/lib/tx-gas-utils.js
@@ -4,7 +4,7 @@ const {
   BnMultiplyByFraction,
   bnToHex,
 } = require('./util')
-const addHexPrefix = require('ethereumjs-util').addHexPrefix
+const { addHexPrefix, isValidAddress } = require('ethereumjs-util')
 const SIMPLE_GAS_COST = '0x5208' // Hex for 21000, cost of a simple send.
 
 /*
@@ -113,12 +113,14 @@ module.exports = class TxGasUtil {
     }
   }
   validateRecipient (txParams) {
-    if (txParams.to === '0x') {
+    if (txParams.to === '0x' || txParams.to === null ) {
       if (txParams.data) {
         delete txParams.to
       } else {
         throw new Error('Invalid recipient address')
       }
+    } else if ( txParams.to !== undefined && !isValidAddress(txParams.to) ) {
+      throw new Error('Invalid recipient address')
     }
     return txParams
   }